+++ maptasker/src/mapit.py
@@ -694,7 +694,7 @@ def mapit_all(file_to_get: str) -> int:
     If coming from the GUI, then PrimeItems may already be primed with data.
     """
 
-    # Save our mapview and doing_diagram flags since 'initialize_everything' would otherwise wipe it out.
+    # Save our mapview and doing_diagram flags since 'initialize_everything' would otherwise wipe them out.
     try:
         save_map = PrimeItems.program_arguments["guiview"]
     except (KeyError, TypeError):