+++ maptasker/src/sysconst.py
@@ -193,6 +193,8 @@ pattern13 = (
 # pattern14 = r"(;Configuration Parameter\(s\):)(.*?)<\\"  # Match everything after the label until a '<'
 pattern14 = r"(;Configuration Parameter\(s\):)(.*?)<span>"
 pattern15 = re.compile("\n")
+# ICON_MAP = {"&#9940;": "⛔", "&#11013;": "⬅️"}
+# pattern16 = re.compile(r"&#9940;|&#11013;")  # "&#9940;": "⛔", "&#11013;": "⬅️"
 
 RE_FONT = re.compile(r"</font>")